+++ controllers/suit_test.go
@@ -20,15 +20,11 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/go-logr/logr"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"go.uber.org/fx"
 
 	"github.com/chaos-mesh/chaos-mesh/api/v1alpha1"
 	"github.com/chaos-mesh/chaos-mesh/controllers/schedule/utils"
-	"github.com/chaos-mesh/chaos-mesh/controllers/types"
-	"github.com/chaos-mesh/chaos-mesh/pkg/workflow/controllers"
 
 	"k8s.io/client-go/rest"
 	"k8s.io/kubectl/pkg/scheme"
@@ -101,19 +97,3 @@ var _ = SynchronizedAfterSuite(func() {}, func() {
 	err := testEnv.Stop()
 	Expect(err).ToNot(HaveOccurred())
 }, 60)
-
-type RunParams struct {
-	fx.In
-
-	Mgr    ctrl.Manager
-	Logger logr.Logger
-
-	Controllers []types.Controller `group:"controller"`
-	Objs        []types.Object     `group:"objs"`
-}
-
-func Run(params RunParams) error {
-	lister = utils.NewActiveLister(k8sClient, params.Logger)
-	err := controllers.BootstrapWorkflowControllers(params.Mgr, params.Logger)
-	return err
-}
